buildvars FIPS: harmonize file naming to match Go stdlib boring convention
Replace the 4-file split (fips.go untagged + fips_on.go tagged + 2 tests)
with a 2-file pair following crypto/internal/boring's pattern:

  fips.go      (//go:build fips)  -> var FIPSEnabled = true
  nofips.go    (//go:build !fips) -> var FIPSEnabled = false
  fips_test.go (//go:build fips)  -> TestFIPSEnabled_TrueUnderFIPSTag
  nofips_test.go (//go:build !fips) -> TestFIPSEnabled_FalseUnderNoFIPSTag

Each compiled build sees exactly one declaration; no init() needed. The
'no<base>' prefix for the complement file matches Go stdlib's
crypto/internal/boring/notboring.go style and reads better than the
prior _on/_off pair.

No call-site changes — buildvars.FIPSEnabled is still a plain bool read.
