+++ buildvars/fips.go
@@ -1,3 +1,5 @@
+//go:build fips
+
 /*
  * SPDX-FileCopyrightText: © 2017-2026 Istari Digital, Inc.
  * SPDX-License-Identifier: Apache-2.0
@@ -6,9 +8,7 @@
 package buildvars
 
 // FIPSEnabled reports whether this binary was built with -tags=fips and
-// is therefore restricted to FIPS 140-3 validated cryptography. Default
-// false; the //go:build fips init() in buildvars/fips_on.go sets it to
-// true at package load before any caller's main() or test body runs.
+// is therefore restricted to FIPS 140-3 validated cryptography.
 //
 // Test code uses it to skip cases the FIPS-tagged binary cannot satisfy:
 //
@@ -16,5 +16,7 @@ package buildvars
 //	    t.Skip("test requires features unavailable under FIPS")
 //	}
 //
-// Read-only after package init.
-var FIPSEnabled = false
+// Companion declaration in buildvars/nofips.go (//go:build !fips) sets
+// the var to false; only one of the two files compiles into any given
+// build, matching Go's stdlib boring/notboring convention.
+var FIPSEnabled = true
