+++ buildvars/fips_test.go
@@ -1,4 +1,4 @@
-//go:build !fips
+//go:build fips
 
 /*
  * SPDX-FileCopyrightText: © 2017-2026 Istari Digital, Inc.
@@ -9,12 +9,11 @@ package buildvars
 
 import "testing"
 
-// TestFIPSEnabled_DefaultFalse confirms the default: without -tags=fips,
-// FIPSEnabled is false. Tagged !fips because the //go:build fips init
-// in buildvars/fips_on.go flips the flag; the FIPS-on assertion lives
-// in the corresponding buildvars/fips_on_test.go (also //go:build fips).
-func TestFIPSEnabled_DefaultFalse(t *testing.T) {
-	if FIPSEnabled {
-		t.Fatal("FIPSEnabled must be false in non-FIPS builds; got true")
+// TestFIPSEnabled_TrueUnderFIPSTag confirms FIPSEnabled is true under
+// -tags=fips (set by buildvars/fips.go's declaration). Companion test
+// in buildvars/nofips_test.go covers the default-false case.
+func TestFIPSEnabled_TrueUnderFIPSTag(t *testing.T) {
+	if !FIPSEnabled {
+		t.Fatal("FIPSEnabled must be true under -tags=fips; got false")
 	}
 }
